Givaro::Caster

+++ linbox/linbox/field/NTL/ntl-RR.h
@@ -8,20 +8,20 @@
  *
  * ------------------------------------
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -288,6 +288,7 @@ namespace LinBox
 		 */
 		Element& random(Element &elt) const
 		{
+			// NTL::random(elt);
 			// Create new random elements
 			if (_size == 0)
 				elt = rand();
@@ -295,7 +296,7 @@ namespace LinBox
 				elt = static_cast<double>((double(rand())/RAND_MAX)*double(_size));
 
 #ifdef TRACE
-			double temp = elt;
+			double temp = (double) elt;
 			std::cout << "random double = " << temp << "    random Element = " << elt << std::endl;
 #endif // TRACE
